+++ org/opensha/sha/imr/attenRelImpl/BJF_1997_AttenRel.java
@@ -152,7 +152,7 @@ public class BJF_1997_AttenRel
    */
   public void setSite(Site site) throws ParameterException {
 
-    vs30Param.setValueIgnoreWarning(site.getParameter(VS30_NAME).getValue());
+    vs30Param.setValueIgnoreWarning((Double)site.getParameter(VS30_NAME).getValue());
     this.site = site;
     setPropagationEffectParams();
 
@@ -172,7 +172,7 @@ public class BJF_1997_AttenRel
     this.site = propEffect.getSite();
     this.eqkRupture = propEffect.getEqkRupture();
 
-    vs30Param.setValueIgnoreWarning(site.getParameter(VS30_NAME).getValue());
+    vs30Param.setValueIgnoreWarning((Double)site.getParameter(VS30_NAME).getValue());
 
     magParam.setValueIgnoreWarning(new Double(eqkRupture.getMag()));
     setFaultTypeFromRake(eqkRupture.getAveRake());
