+++ scratchJavaDevelopers/jennie/attenRelImpl/ToroEtAl_1997_SiteSpecific_AttenRel2.java
@@ -223,12 +223,12 @@ public class ToroEtAl_1997_SiteSpecific_AttenRel2
   public void setSite(Site site) throws ParameterException {
 
 	  AF_FuncForm.setValue(AF_FuncForm_DEFAULT);
-	  AF_InterceptParam.setValue(site.getParameter(AF_INTERCEPT_PARAM_NAME).getValue());
-	  AF_AddRefAccParam.setValue(site.getParameter(AF_ADDITIVE_REF_ACCELERATION_PARAM_NAME).getValue());
-	  AF_SlopeParam.setValue(site.getParameter(AF_SLOPE_PARAM_NAME).getValue());
-	  AF_MagParam.setValue(site.getParameter(AF_MagPARAM_NAME).getValue());
-	  AF_RupParam.setValue(site.getParameter(AF_RupPARAM_NAME).getValue());
-	  AF_StdDevParam.setValue(site.getParameter(AF_STD_DEV_PARAM_NAME).getValue());  
+	  AF_InterceptParam.setValue((Double)site.getParameter(AF_INTERCEPT_PARAM_NAME).getValue());
+	  AF_AddRefAccParam.setValue((Double)site.getParameter(AF_ADDITIVE_REF_ACCELERATION_PARAM_NAME).getValue());
+	  AF_SlopeParam.setValue((Double)site.getParameter(AF_SLOPE_PARAM_NAME).getValue());
+	  AF_MagParam.setValue((Double)site.getParameter(AF_MagPARAM_NAME).getValue());
+	  AF_RupParam.setValue((Double)site.getParameter(AF_RupPARAM_NAME).getValue());
+	  AF_StdDevParam.setValue((Double)site.getParameter(AF_STD_DEV_PARAM_NAME).getValue());  
 	  this.site = site;
 	  // set the location in as_2008_attenRel
 	  toroEtAl_1997_attenRel.setSiteLocation(site.getLocation());
